[feature] Implement my orders list and change layout (#35)
* [feature] Implement my orders list and change layout

* [bugfix] Make sure approval is verified for selected input token (even if balance = 0)

* [performance] Centralize orderbook synchronization

create a separate compononet for managing orderbook refresh
This is to avoid calling 0x API everytime Riot component is update
and instead controlling it by having a single call each second

* [bugfix] Remove reference to old undeclared orderBook object

* [bugfix] Fix order cancel failure

this was due to missing import. Did refactor
the common imports to keep related functions
in the same modules

* [layout] Correct token swap menu background color

* [ux] Only autofocus orderbook at first display

When scrolling through the orderbook. Autofocus,
that triggers at each orderbook update, is causing
the scroll to refocuse to the middle of the orderbook.

This makes it impossible to manually scroll the orderbook
So changing this disable autofocuse when user starts
scrolling through the orderbook

* [feature] Implement new limit order dashboard

* [bugfix] List all order types independently of type of order selection

* [ux] Check approval is needed only at order submit time

As approval check is asynchronous and can take some time
It is causing ux issues if checked everytime user chooses
a new token or changes values. Changed this to only
checking approval when actually requesting the order submit
as it is the only relevant event for checking approvals

* [refactor] refactor fee logic into dedicated module

* [bugfix] Correct order matching logic

* [bugfix] Display remaining unfilled volumes in order book

* [bugfix] Update token selector component after close action

When applying a filter and selecting the token or closing
the list is not updated.

* [bugfix] Correct order partial match logic

* [feature] Limit exchanged volumes to > 10 DAI

* [feature] Add mid market price in orderbook

* [feature] Display my order filled DAI amount

* [feature] Add HYPE to static token list

* [bugfix] Correct amount calculations to always return integer values

* [cleanup] Remove old unused css file

* [bugfix] Correct order book amount displays

Taker/Maker amounts were calculated with the wrong
decimal values.

* [bugfix] Handle empty reply when callig Coingecko price api

* [feature] Do not submit orders with value <= 10 DAI

* [feature] Disable submit button when oder <= 10

* [feature] Disable click pointer on inactive submit button

* [feature] Set gas price to fast when submitting orders

* [bufix] Better handle gas price oracle API failure

* [refactor] refactor common API calls + reorganize project folders

* [layout] Correct token selector colors

* [layout] Change default token logo when theme changes

* [bugfix] Correct token price query

* [feature] Disable DAI from input token list when in limit order mode

* [ux] Disable order form amount/price input formatting for better usability

* [ux] Add loader when orderbook is being updated after token selection

* [ux] Add loader to my order list when token selection changes

* [layout] Adjust orderbook canvas height for better display

* [ux] Keep filtered tokens in dom to avoid constant reload of logo images

* [bugfix] Fix order amount display

* [bugfix] Make sure to display account address when available

* [bugfix] While checking approval request approval for appropriate token

Depending on the order type, input token approval should be requested
if it is a sell, output token otherwise